Moving all Exception declarations to errors.py
