+++ datacats/environment.py
@@ -21,14 +21,14 @@ from ConfigParser import (SafeConfigParser, Error as ConfigParserError,
 from datacats.validate import valid_name
 from datacats.docker import (web_command, run_container, remove_container,
                              inspect_container, is_boot2docker, data_only_container, docker_host,
-                             PortAllocatedError, container_logs, remove_image, WebCommandError,
+                             container_logs, remove_image,
                              image_exists)
 from datacats.template import ckan_extension_template
 from datacats.scripts import (WEB, SHELL, PASTER, PASTER_CD, PURGE,
     RUN_AS_USER, INSTALL_REQS, CLEAN_VIRTUALENV, INSTALL_PACKAGE,
     COMPILE_LESS)
 from datacats.network import wait_for_service_available, ServiceTimeout
-from datacats.error import DatacatsError
+from datacats.error import DatacatsError, WebCommandError, PortAllocatedError
 
 
 WEB_START_TIMEOUT_SECONDS = 30
